Add basic app to test angular
I did this because angular gets into a crazy looping state when it loads
the application. This seems to only be caused by including office JS and
angular and how they interact.

What appears to be happening is that the `href` (as seen when calling
`window.location.href` in the inspector after the task pane add-in is
sideloaded) is infinitely appending
`?_host_Info=Word$Mac$16.01$en-US$$$$0#/` to the url. This eventually
causes processing resource issues on the computer. It doesn't appear to
be a true redirect because the `#` is not triggering a reload in the
browser because it's still the same url.

This seems to be an issue with the combination of Office JS and Angular
as I have not seen the issue when loading a task pane add in from a
simple HTML file that loads the Office.js script.

Hash location strategy is not on in this commit. They suggest using it
but it seems like this isn't a routing problem but an issue with how the
JS is manipulating the history push state or window location href.

https://docs.microsoft.com/en-us/office/dev/add-ins/develop/add-ins-with-angular2#use-the-hash-location-strategy-in-the-angular-application