HashLocationStrategy set to true
With this setting on, the problem doesn't happen. This doesn't seem like
the best solution but understand if it's the only way.

I think this "fixes" the problem because angular will load `index` from
`https://localhost:4200/` and ignore the query params of
`?_host_Info=Word$Mac$16.01$en-US$$$$0` and handle the routing via
angular because of the added `#/` at the end.

Here is an example of the whole url when calling `window.location.href`

```
https://localhost:4200/?_host_Info=Word$Mac$16.01$en-US$$$$0#/
```