Move office js script to loading later in the angular lifecycle
I did this because when the JS is loaded later in the application life
cycle, it does not create the href issue and Office appears to load
correctly.

I think this is working because the index is loaded, the routing happens
by angular and then office JS is loaded, which manipulates the dom by
loading several scripts where it eventually resolves the
`Office.onReady()` promise.